requirements.txt
streamlit
pillow
torch
torchvision
scikit-learn 